+++ CHANGELOG.md
@@ -5,13 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/wickeyware/wickeyappstore/compare/2.11.0...HEAD)
+## [Unreleased](https://github.com/wickeyware/wickeyappstore/compare/2.12.1...HEAD)
 
 This shows all commits since last release (actually, currently all commits).
 
+## [2.12.1](https://github.com/wickeyware/wickeyappstore/compare/2.12.0...2.12.1) - 2018-05-31
+
+### Fixed
+
+* **WASjs** Remove ViewEncapsulation.Native (a polyfill was required for all browsers other than Chrome).
+* **WasDataService** Add .get(_key) that returns value not observable.
+* **package** Update to latest patch releases.
+* **demo** Fix WickeyAppStore import in the demo app.
+
 ## [2.12.0](https://github.com/wickeyware/wickeyappstore/compare/2.11.1...2.12.0) - 2018-05-30
 
-### Added 
+### Added
 
 * **WASMenu** Add hook to indicate when WAS is opened. <wickey-appstore (open)="pause()"></wickey-appstore>.